+++ sdk/core/azure-core/tests/test_pipeline.py
@@ -434,5 +434,17 @@ class TestClientRequest(unittest.TestCase):
         with pytest.raises(ValueError):
             client = PipelineClient(base_url="test", policies=policies, per_retry_policies=[foo_policy])
 
+    def test_sansiohttppolicy_exception(self):
+        class ReproPolicy(SansIOHTTPPolicy):
+            def on_exception(self, request):
+                return True
+
+        class ErrorPolicy(HTTPPolicy):
+            def send(self, request):
+                raise Exception("oops")
+
+        pipeline = Pipeline(transport=object(), policies=[ReproPolicy(), ErrorPolicy()])
+        pipeline.run(HttpRequest("GET", "..."))
+
 if __name__ == "__main__":
     unittest.main()